# Conflicts:
#	community/sources/webcore/src/canopsis-next/src/components/modals/index.js

+++ community/go-engines-community/lib/api/scenario/api.go
@@ -3,6 +3,8 @@ package scenario
 import (
 	"context"
 	"encoding/json"
+	"net/http"
+
 	"git.canopsis.net/canopsis/canopsis-community/community/go-engines-community/lib/api/auth"
 	"git.canopsis.net/canopsis/canopsis-community/community/go-engines-community/lib/api/common"
 	"git.canopsis.net/canopsis/canopsis-community/community/go-engines-community/lib/api/logger"
@@ -11,12 +13,13 @@ import (
 	"github.com/gin-gonic/gin"
 	"github.com/gin-gonic/gin/binding"
 	"github.com/valyala/fastjson"
-	"net/http"
 )
 
 type api struct {
-	store             Store
-	actionLogger      logger.ActionLogger
+	store        Store
+	actionLogger logger.ActionLogger
+
+	//todo: priority intervals with new requirements are looks weird now, should think about cleaner solution
 	priorityIntervals action.PriorityIntervals
 }
 
@@ -38,23 +41,8 @@ func NewApi(
 	}
 }
 
-// Find all scenarios
-// @Summary Find scenarios
-// @Description Get paginated list of scenarios
-// @Tags scenarios
-// @ID scenarios-find-all
-// @Accept json
-// @Produce json
-// @Security ApiKeyAuth
-// @Security BasicAuth
-// @Param page query integer true "current page"
-// @Param limit query integer true "items per page"
-// @Param search query string false "search query"
-// @Param sort query string false "sort query"
-// @Param sort_by query string false "sort query"
+// List
 // @Success 200 {object} common.PaginatedListResponse{data=[]Scenario}
-// @Failure 400 {object} common.ValidationErrorResponse
-// @Router /scenarios [get]
 func (a *api) List(c *gin.Context) {
 	var query FilteredQuery
 	query.Query = pagination.GetDefaultQuery()
@@ -78,18 +66,8 @@ func (a *api) List(c *gin.Context) {
 	c.JSON(http.StatusOK, res)
 }
 
-// Get scenario by id
-// @Summary Get scenario by id
-// @Description Get scenario by id
-// @Tags scenarios
-// @ID scenarios-get-by-id
-// @Produce json
-// @Security ApiKeyAuth
-// @Security BasicAuth
-// @Param id path string true "scenario id"
+// Get
 // @Success 200 {object} Scenario
-// @Failure 404 {object} common.ErrorResponse
-// @Router /scenarios/{id} [get]
 func (a *api) Get(c *gin.Context) {
 	scenario, err := a.store.GetOneBy(c.Request.Context(), c.Param("id"))
 	if err != nil {
@@ -103,21 +81,14 @@ func (a *api) Get(c *gin.Context) {
 	c.JSON(http.StatusOK, scenario)
 }
 
-// Create scenario
-// @Summary Create scenario
-// @Description Create scenario
-// @Tags scenarios
-// @ID scenarios-create
-// @Accept json
-// @Produce json
-// @Security ApiKeyAuth
-// @Security BasicAuth
+// Create
 // @Param body body EditRequest true "body"
 // @Success 201 {object} Scenario
-// @Failure 400 {object} common.ValidationErrorResponse
-// @Router /scenarios [post]
 func (a *api) Create(c *gin.Context) {
 	var request CreateRequest
+
+	priority := a.priorityIntervals.GetMinimal()
+	request.Priority = &priority
 
 	if err := c.ShouldBind(&request); err != nil {
 		c.JSON(http.StatusBadRequest, common.NewValidationErrorResponse(err, request))
@@ -125,14 +96,15 @@ func (a *api) Create(c *gin.Context) {
 	}
 
 	userId := c.MustGet(auth.UserKey).(string)
-	author := c.MustGet(auth.Username).(string)
-	setActionParameterAuthorAndUserID(&request.EditRequest, author, userId)
 
 	scenario, err := a.store.Insert(c.Request.Context(), request)
 	if err != nil {
 		panic(err)
 	}
-
+	if scenario == nil {
+		c.AbortWithStatusJSON(http.StatusNotFound, common.NotFoundResponse)
+		return
+	}
 	a.priorityIntervals.Take(scenario.Priority)
 
 	err = a.actionLogger.Action(context.Background(), userId, logger.LogEntry{
@@ -147,21 +119,9 @@ func (a *api) Create(c *gin.Context) {
 	c.JSON(http.StatusCreated, scenario)
 }
 
-// Update scenario by id
-// @Summary Update scenario by id
-// @Description Update scenario by id
-// @Tags scenarios
-// @ID scenarios-update-by-id
-// @Accept json
-// @Produce json
-// @Security ApiKeyAuth
-// @Security BasicAuth
-// @Param id path string true "scenario id"
+// Update
 // @Param body body EditRequest true "body"
 // @Success 200 {object} Scenario
-// @Failure 400 {object} common.ValidationErrorResponse
-// @Failure 404 {object} common.ErrorResponse
-// @Router /scenarios/{id} [put]
 func (a *api) Update(c *gin.Context) {
 	request := UpdateRequest{
 		ID: c.Param("id"),
@@ -176,14 +136,15 @@ func (a *api) Update(c *gin.Context) {
 		return
 	}
 
+	priority := a.priorityIntervals.GetMinimal()
+	request.Priority = &priority
+
 	if err := c.ShouldBind(&request); err != nil {
 		c.JSON(http.StatusBadRequest, common.NewValidationErrorResponse(err, request))
 		return
 	}
 
 	userId := c.MustGet(auth.UserKey).(string)
-	author := c.MustGet(auth.Username).(string)
-	setActionParameterAuthorAndUserID(&request.EditRequest, author, userId)
 
 	newScenario, err := a.store.Update(c.Request.Context(), request)
 	if err != nil {
@@ -209,17 +170,6 @@ func (a *api) Update(c *gin.Context) {
 	c.JSON(http.StatusOK, newScenario)
 }
 
-// Delete scenario by id
-// @Summary Delete scenario by id
-// @Description Delete scenario by id
-// @Tags scenarios
-// @ID scenarios-delete-by-id
-// @Security ApiKeyAuth
-// @Security BasicAuth
-// @Param id path string true "scenario id"
-// @Success 204
-// @Failure 404 {object} common.ErrorResponse
-// @Router /scenarios/{id} [delete]
 func (a *api) Delete(c *gin.Context) {
 	id := c.Param("id")
 
@@ -257,35 +207,17 @@ func (a *api) Delete(c *gin.Context) {
 	c.Status(http.StatusNoContent)
 }
 
-// Get minimal priority
-// @Summary Get minimal priority
-// @Description Get minimal priority
-// @Tags scenarios
-// @ID scenarios-get-minimal-priority
-// @Produce json
-// @Security ApiKeyAuth
-// @Security BasicAuth
+// GetMinimalPriority
 // @Success 200 {object} GetMinimalPriorityResponse
-// @Router /scenarios/minimal-priority [get]
 func (a *api) GetMinimalPriority(c *gin.Context) {
 	c.JSON(http.StatusOK, GetMinimalPriorityResponse{
 		Priority: a.priorityIntervals.GetMinimal(),
 	})
 }
 
-// Check priority
-// @Summary Check priority
-// @Description Check priority
-// @Tags scenarios
-// @ID scenarios-check-priority
-// @Accept json
-// @Produce json
-// @Security ApiKeyAuth
-// @Security BasicAuth
+// CheckPriority
 // @Param body body CheckPriorityRequest true "body"
 // @Success 200 {object} CheckPriorityResponse
-// @Failure 400 {object} common.ValidationErrorResponse
-// @Router /scenarios/check-priority [post]
 func (a *api) CheckPriority(c *gin.Context) {
 	request := CheckPriorityRequest{}
 	if err := c.ShouldBind(&request); err != nil {
@@ -309,19 +241,9 @@ func (a *api) CheckPriority(c *gin.Context) {
 	})
 }
 
-// Bulk create scenarios
-// @Summary Bulk create scenarios
-// @Description Bulk create scenarios
-// @Tags scenarios
-// @ID scenarios-bulk-create
-// @Accept json
-// @Produce json
-// @Security JWTAuth
-// @Security BasicAuth
+// BulkCreate
 // @Param body body []CreateRequest true "body"
 // @Success 207 {array} []BulkCreateResponseItem
-// @Failure 400 {object} common.ValidationErrorResponse
-// @Router /bulk/scenarios [post]
 func (a *api) BulkCreate(c *gin.Context) {
 	userId := c.MustGet(auth.UserKey).(string)
 
@@ -361,13 +283,16 @@ func (a *api) BulkCreate(c *gin.Context) {
 			continue
 		}
 
+		if request.Priority == nil {
+			priority := a.priorityIntervals.GetMinimal()
+			request.Priority = &priority
+		}
+
 		err = binding.Validator.ValidateStruct(request)
 		if err != nil {
 			response.SetArrayItem(idx, common.GetBulkResponseItem(&ar, "", http.StatusBadRequest, rawObject, common.NewValidationErrorFastJsonValue(&ar, err, request)))
 			continue
 		}
-
-		setActionParameterAuthorAndUserID(&request.EditRequest, c.MustGet(auth.Username).(string), userId)
 
 		scenario, err := a.store.Insert(ctx, request)
 		if err != nil {
@@ -375,6 +300,7 @@ func (a *api) BulkCreate(c *gin.Context) {
 			continue
 		}
 
+		a.priorityIntervals.Take(scenario.Priority)
 		response.SetArrayItem(idx, common.GetBulkResponseItem(&ar, scenario.ID, http.StatusOK, rawObject, nil))
 
 		err = a.actionLogger.Action(context.Background(), userId, logger.LogEntry{
@@ -390,19 +316,9 @@ func (a *api) BulkCreate(c *gin.Context) {
 	c.Data(http.StatusMultiStatus, gin.MIMEJSON, response.MarshalTo(nil))
 }
 
-// Bulk update scenarios
-// @Summary Bulk update scenarios
-// @Description Bulk update scenarios
-// @Tags scenarios
-// @ID scenarios-bulk-update
-// @Accept json
-// @Produce json
-// @Security JWTAuth
-// @Security BasicAuth
+// BulkUpdate
 // @Param body body []BulkUpdateRequestItem true "body"
 // @Success 207 {array} []BulkUpdateResponseItem
-// @Failure 400 {object} common.ValidationErrorResponse
-// @Router /bulk/scenarios [put]
 func (a *api) BulkUpdate(c *gin.Context) {
 	userId := c.MustGet(auth.UserKey).(string)
 
@@ -442,13 +358,27 @@ func (a *api) BulkUpdate(c *gin.Context) {
 			continue
 		}
 
+		if request.Priority == nil {
+			priority := a.priorityIntervals.GetMinimal()
+			request.Priority = &priority
+		}
+
 		err = binding.Validator.ValidateStruct(request)
 		if err != nil {
 			response.SetArrayItem(idx, common.GetBulkResponseItem(&ar, "", http.StatusBadRequest, rawObject, common.NewValidationErrorFastJsonValue(&ar, err, request)))
 			continue
 		}
 
-		setActionParameterAuthorAndUserID(&request.EditRequest, c.MustGet(auth.Username).(string), userId)
+		oldScenario, err := a.store.GetOneBy(c.Request.Context(), request.ID)
+		if err != nil {
+			response.SetArrayItem(idx, common.GetBulkResponseItem(&ar, "", http.StatusInternalServerError, rawObject, ar.NewString(err.Error())))
+			continue
+		}
+
+		if oldScenario == nil {
+			response.SetArrayItem(idx, common.GetBulkResponseItem(&ar, "", http.StatusNotFound, rawObject, ar.NewString("Not found")))
+			continue
+		}
 
 		scenario, err := a.store.Update(ctx, UpdateRequest(request))
 		if err != nil {
@@ -461,6 +391,8 @@ func (a *api) BulkUpdate(c *gin.Context) {
 			continue
 		}
 
+		a.priorityIntervals.Restore(oldScenario.Priority)
+		a.priorityIntervals.Take(scenario.Priority)
 		response.SetArrayItem(idx, common.GetBulkResponseItem(&ar, scenario.ID, http.StatusOK, rawObject, nil))
 
 		err = a.actionLogger.Action(context.Background(), userId, logger.LogEntry{
@@ -476,19 +408,9 @@ func (a *api) BulkUpdate(c *gin.Context) {
 	c.Data(http.StatusMultiStatus, gin.MIMEJSON, response.MarshalTo(nil))
 }
 
-// Bulk delete scenarios
-// @Summary Bulk delete scenarios
-// @Description Bulk delete scenarios
-// @Tags scenarios
-// @ID scenarios-bulk-delete
-// @Accept json
-// @Produce json
-// @Security JWTAuth
-// @Security BasicAuth
+// BulkDelete
 // @Param body body []BulkDeleteRequestItem true "body"
 // @Success 207 {array} []BulkDeleteResponseItem
-// @Failure 400 {object} common.ValidationErrorResponse
-// @Router /bulk/scenarios [delete]
 func (a *api) BulkDelete(c *gin.Context) {
 	userId := c.MustGet(auth.UserKey).(string)
 
@@ -534,6 +456,17 @@ func (a *api) BulkDelete(c *gin.Context) {
 			continue
 		}
 
+		scenario, err := a.store.GetOneBy(c.Request.Context(), request.ID)
+		if err != nil {
+			response.SetArrayItem(idx, common.GetBulkResponseItem(&ar, "", http.StatusInternalServerError, rawObject, ar.NewString(err.Error())))
+			continue
+		}
+
+		if scenario == nil {
+			response.SetArrayItem(idx, common.GetBulkResponseItem(&ar, "", http.StatusNotFound, rawObject, ar.NewString("Not found")))
+			continue
+		}
+
 		ok, err := a.store.Delete(ctx, request.ID)
 		if err != nil {
 			response.SetArrayItem(idx, common.GetBulkResponseItem(&ar, "", http.StatusInternalServerError, rawObject, ar.NewString(err.Error())))
@@ -545,6 +478,7 @@ func (a *api) BulkDelete(c *gin.Context) {
 			continue
 		}
 
+		a.priorityIntervals.Restore(scenario.Priority)
 		response.SetArrayItem(idx, common.GetBulkResponseItem(&ar, request.ID, http.StatusOK, rawObject, nil))
 
 		err = a.actionLogger.Action(context.Background(), userId, logger.LogEntry{
@@ -558,31 +492,4 @@ func (a *api) BulkDelete(c *gin.Context) {
 	}
 
 	c.Data(http.StatusMultiStatus, gin.MIMEJSON, response.MarshalTo(nil))
-}
-
-func setActionParameterAuthorAndUserID(request *EditRequest, author, userID string) {
-	for i, action := range request.Actions {
-		switch v := action.Parameters.(type) {
-		case SnoozeParametersRequest:
-			v.Author = author
-			v.User = userID
-			request.Actions[i].Parameters = v
-		case ChangeStateParametersRequest:
-			v.Author = author
-			v.User = userID
-			request.Actions[i].Parameters = v
-		case AssocTicketParametersRequest:
-			v.Author = author
-			v.User = userID
-			request.Actions[i].Parameters = v
-		case PbehaviorParametersRequest:
-			v.Author = author
-			v.User = userID
-			request.Actions[i].Parameters = v
-		case ParametersRequest:
-			v.Author = author
-			v.User = userID
-			request.Actions[i].Parameters = v
-		}
-	}
 }
